+++ backend/canisters/community/api/src/updates/c2c_send_moderation_report.rs
@@ -1,0 +1,10 @@
+use serde::{Deserialize, Serialize};
+use types::{ChannelId, UnitResult};
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct Args {
+    pub channel_id: ChannelId,
+    pub text: String,
+}
+
+pub type Response = UnitResult;