+++ backend/canisters/local_user_index/impl/src/queries/access_token_v2.rs
@@ -11,7 +11,8 @@ use types::c2c_can_issue_access_token::{
     AccessTypeArgs, BotActionByCommandArgs, JoinVideoCallArgs, MarkVideoCallAsEndedArgs, StartVideoCallArgs,
 };
 use types::{
-    AutonomousBotScope, BotActionByCommandClaims, BotCommand, Chat, JoinOrEndVideoCallClaims, Milliseconds,
+    AutonomousBotScope, BotActionByCommandClaims, BotCommand, CLAIM_TYPE_BOT_ACTION_BY_COMMAND, CLAIM_TYPE_JOIN_VIDEO_CALL,
+    CLAIM_TYPE_MARK_VIDEO_CALL_AS_ENDED, CLAIM_TYPE_START_VIDEO_CALL, Chat, JoinOrEndVideoCallClaims, Milliseconds,
     StartVideoCallClaims, TranslateClaims,
 };
 
@@ -222,10 +223,10 @@ impl ArgsInternal {
 
     pub fn type_name(&self) -> &str {
         match self {
-            Self::StartVideoCall(_) => "StartVideoCall",
-            Self::JoinVideoCall(_) => "JoinVideoCall",
-            Self::MarkVideoCallAsEnded(_) => "MarkVideoCallAsEnded",
-            Self::BotActionByCommand(_) => "BotActionByCommand",
+            Self::StartVideoCall(_) => CLAIM_TYPE_START_VIDEO_CALL,
+            Self::JoinVideoCall(_) => CLAIM_TYPE_JOIN_VIDEO_CALL,
+            Self::MarkVideoCallAsEnded(_) => CLAIM_TYPE_MARK_VIDEO_CALL_AS_ENDED,
+            Self::BotActionByCommand(_) => CLAIM_TYPE_BOT_ACTION_BY_COMMAND,
             Self::Translate => "Translate",
         }
     }