+++ frontend/app/src/components_mobile/home/DeletedContent.svelte
@@ -51,7 +51,6 @@
                         username,
                         timestamp: timestampStr,
                         rules: "/guidelines?section=3",
-                        modclub: "https://modclub.ai/",
                     },
                 })} />
         {:else}